# docs/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: LeetGaming Pro API
  description: |
    # 🎮 LeetGaming Pro - Competitive Gaming Platform API

    Welcome to the LeetGaming Pro API documentation. This API powers a professional-grade
    competitive gaming platform with features including:

    - **Authentication & Security**: OAuth2, MFA/TOTP, JWT tokens
    - **Matchmaking**: Glicko-2 rating system, skill-based matching
    - **Tournaments**: Bracket generation, prize pools, blockchain verification
    - **Replays**: Upload, analysis, sharing with demo parsing
    - **Squads**: Team management, invitations, statistics
    - **Billing**: Subscriptions, withdrawals, audit trails

    ## 🔐 Authentication

    Most endpoints require authentication via Bearer token:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## 📊 Rate Limits

    | Tier | Requests/min | Burst |
    |------|--------------|-------|
    | Free | 60 | 10 |
    | Pro | 300 | 50 |
    | Enterprise | 1000 | 200 |

    ## 🌐 Base URLs

    | Environment | URL |
    |-------------|-----|
    | Production | https://api.leetgaming.gg |
    | Staging | https://api.staging.leetgaming.gg |
    | Development | http://localhost:8080 |

  version: 2.0.0
  contact:
    name: LeetGaming Pro Support
    url: https://leetgaming.gg/support
    email: api@leetgaming.gg
  license:
    name: Proprietary
    url: https://leetgaming.gg/terms
servers:
  - url: https://api.leetgaming.gg
    description: Production
...
    description: Staging
...
paths:
...
  /matches/player/{player_id}:
    get:
      tags: [Matches]
      summary: Get Player by ID
      operationId: get_matches_player_player_id
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /matches/squad/{squad_id}:
    get:
      tags: [Matches]
      summary: Get Squad by ID
      operationId: get_matches_squad_squad_id
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /prize-pools/{id}:
    get:
      tags: [Prize Pools]
      summary: Get Prize-Pools by ID
      operationId: get_prize_pools_id
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /prize-pools/{id}/history:
    get:
      tags: [Prize Pools]
      summary: Get History by ID
      operationId: get_prize_pools_id_history
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /matches/{match_id}/prize-pool:
    get:
      tags: [Matches]
      summary: Get Prize-Pool by ID
      operationId: get_matches_match_id_prize_pool
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
...
security:
  - bearerAuth: []